Added cookies for the login credintials, and removed session params
It is a practice on how to use Java Cookies, and how to write a good commit
messages. Since sessions are saved on the server (so the rely on the server),
I switched to cookies to keep the state for longer time, and to keep the
server stateless in terms of user session.